requirements.txt
# prod
alpaca-trade-api==1.2.3
yfinance==0.1.63
python-dotenv==0.18.0
loguru==0.5.3
requests==2.26.0
requests-cache==0.7.1
stockstats==0.3.2
tabulate==0.8.9
finviz==1.4.2
setuptools==57.2.0
setuptools-scm==6.0.1
wheel==0.36.2

# backtesting
backtesting==0.3.1
# matplotlib==3.2.2
